pyproject.toml
[project]
name = "agentic-ai-one"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.10",
    "langchain-groq>=1.1.2",
    "langchain-openai>=1.1.9",
    "langgraph>=1.0.8",
    "python-dotenv>=1.2.1",
]
